tools/requirements.txt
# Requirements automatically generated by pigar.
# https://github.com/Damnever/pigar

# tools/png2kicad_mod.py: 21
# tools/png2kicad_pcb.py: 21
# tools/png2lbr.py: 21
Pillow == 4.3.0

# tools/png2kicad_mod.py: 24
# tools/png2kicad_pcb.py: 24
# tools/png2lbr.py: 24
Shapely == 1.6.2.post1

# tools/png2kicad_mod.py: 22
# tools/png2kicad_pcb.py: 22
# tools/png2lbr.py: 22
numpy == 1.13.3

# tools/png2kicad_mod.py: 23
# tools/png2kicad_pcb.py: 23
# tools/png2lbr.py: 23
pypotrace == 0.1.3
